+++ src/main/Recorder.ts
@@ -496,6 +496,8 @@ export default class Recorder {
       );
     } else if (obsCaptureMode === 'game_capture') {
       this.videoSource = Recorder.createGameCaptureSource(captureCursor);
+    } else if (obsCaptureMode === 'window_capture') {
+      this.videoSource = Recorder.createWindowCaptureSource(captureCursor);
     } else {
       throw new Error(`[Recorder] Unexpected mode: ${obsCaptureMode}`);
     }
@@ -559,6 +561,28 @@ export default class Recorder {
     gameCaptureSource.save();
 
     return gameCaptureSource;
+  }
+
+  /**
+   * Creates a window capture source.
+   */
+  private static createWindowCaptureSource(captureCursor: boolean) {
+    console.info('[Recorder] Configuring OBS for Window Capture');
+
+    const windowCaptureSource = osn.InputFactory.create(
+      'window_capture',
+      'WR Window Capture',
+      {
+        cursor: captureCursor,
+        window: 'World of Warcraft:GxWindowClass:Wow.exe',
+        // This corresponds to Windows Graphics Capture. The other mode "BITBLT" doesn't seem to work and
+        // capture behind the WoW window. Not sure why, some googling suggested Windows theme issues.
+        // See https://github.com/obsproject/obs-studio/blob/master/plugins/win-capture/window-capture.c#L70.
+        method: 2,
+      }
+    );
+
+    return windowCaptureSource;
   }
 
   /**