+++ docs/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add the 3840 x 1200 resolution.
 - [Issue 125](https://github.com/aza547/wow-recorder/issues/125) - Push to talk for microphone recording.
+- Window capture support.
 
 ### Fixed
 - [Issue 443](https://github.com/aza547/wow-recorder/issues/443) - Fix to Ctrl + R breaking the app.